Change structure project
